# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RentCar API
  version: 1.0.0
  description: Monolithic car rental API with user, car, rental and payment modules.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Rotate refresh token and issue new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke one refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Logged out
  /cars:
    get:
      tags: [Cars]
      security: []
      summary: List cars with filters
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: q
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema:
            type: string
            enum: [available, maintenance, inactive]
        - name: min_year
          in: query
          schema: { type: integer }
        - name: max_year
          in: query
          schema: { type: integer }
        - name: min_rate
          in: query
          schema: { type: number }
        - name: max_rate
          in: query
          schema: { type: number }
      responses:
        '200':
          description: Cars
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCarsEnvelope'
  /cars/{id}:
    get:
      tags: [Cars]
      security: []
      summary: Get car by id
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Car
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarEnvelope'
  /availability/cars/{id}:
    get:
      tags: [Rentals]
      security: []
      summary: Check car availability for a date range
      parameters:
        - $ref: '#/components/parameters/ID'
        - name: start_date
          in: query
          required: true
          schema: { type: string, format: date }
        - name: end_date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Availability result
  /availability/cars/{id}/calendar:
    get:
      tags: [Rentals]
      security: []
      summary: Get blocked booking and maintenance days for a car
      parameters:
        - $ref: '#/components/parameters/ID'
        - name: month
          in: query
          schema: { type: string, example: '2026-05' }
      responses:
        '200':
          description: Availability calendar
  /rentals:
    post:
      tags: [Rentals]
      summary: Create rental
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [car_id, start_date, end_date]
              properties:
                car_id: { type: integer, format: int64 }
                start_date: { type: string, format: date }
                end_date: { type: string, format: date }
      responses:
        '201':
          description: Rental created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalEnvelope'
  /rentals/me:
    get:
      tags: [Rentals]
      summary: List current user's rentals
      responses:
        '200':
          description: Rentals
  /rentals/{id}:
    get:
      tags: [Rentals]
      summary: Get rental if owner or admin
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Rental
  /rentals/{id}/cancel:
    post:
      tags: [Rentals]
      summary: Cancel eligible rental
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Cancelled
  /rentals/{id}/invoice:
    get:
      tags: [Invoices]
      summary: Get or create rental invoice
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Invoice
  /notifications:
    get:
      tags: [Notifications]
      summary: List current user's notifications
      parameters:
        - name: unread
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Notifications
  /notifications/{id}/read:
    post:
      tags: [Notifications]
      summary: Mark notification as read
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Marked read
  /payments:
    post:
      tags: [Payments]
      summary: Create payment request for pending rental
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rental_id, method]
              properties:
                rental_id: { type: integer, format: int64 }
                method:
                  type: string
                  enum: [cash, card, bank_transfer]
      responses:
        '201':
          description: Payment created
  /users/me:
    get:
      tags: [Users]
      summary: Current user profile
      responses:
        '200':
          description: User
    patch:
      tags: [Users]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name: { type: string }
                email: { type: string, format: email }
      responses:
        '200':
          description: Updated user
  /admin/cars:
    post:
      tags: [Admin Cars]
      summary: Create car
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarInput'
      responses:
        '201':
          description: Created
  /admin/cars/{id}:
    put:
      tags: [Admin Cars]
      summary: Update car
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarInput'
      responses:
        '200':
          description: Updated
    delete:
      tags: [Admin Cars]
      summary: Delete car
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Deleted
  /admin/rentals:
    get:
      tags: [Admin Rentals]
      summary: List all rentals with filters
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            type: string
            enum: [requested, approved, rejected, pending_payment, confirmed, active, cancelled, completed]
        - name: payment_status
          in: query
          schema:
            type: string
            enum: [pending, paid, failed, refunded]
        - name: user_id
          in: query
          schema: { type: integer, format: int64 }
        - name: car_id
          in: query
          schema: { type: integer, format: int64 }
        - name: start_from
          in: query
          schema: { type: string, format: date }
        - name: end_to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Paginated rentals
  /admin/rentals/{id}/status:
    patch:
      tags: [Admin Rentals]
      summary: Update rental status
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [requested, approved, rejected, pending_payment, confirmed, active, cancelled, completed]
      responses:
        '200':
          description: Updated
  /admin/rentals/{id}/approve:
    post:
      tags: [Admin Rentals]
      summary: Approve requested rental
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Approved
  /admin/rentals/{id}/reject:
    post:
      tags: [Admin Rentals]
      summary: Reject requested rental
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Rejected
  /admin/payments/{id}/confirm:
    post:
      tags: [Admin Payments]
      summary: Confirm pending payment
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Confirmed
  /admin/payments/{id}/fail:
    post:
      tags: [Admin Payments]
      summary: Reject pending payment
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Failed
  /admin/payments/{id}/refund:
    post:
      tags: [Admin Payments]
      summary: Refund paid payment and cancel rental
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Refunded
  /admin/users:
    get:
      tags: [Admin Users]
      summary: List users; super admin only
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: q
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema:
            type: string
            enum: [customer, admin, super_admin]
      responses:
        '200':
          description: Paginated users
  /admin/maintenance:
    get:
      tags: [Admin Maintenance]
      summary: List maintenance windows
      responses:
        '200':
          description: Maintenance records
    post:
      tags: [Admin Maintenance]
      summary: Create maintenance window
      responses:
        '201':
          description: Created
  /admin/maintenance/{id}:
    put:
      tags: [Admin Maintenance]
      summary: Update maintenance window
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Updated
    delete:
      tags: [Admin Maintenance]
      summary: Delete maintenance window
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Deleted
  /admin/reports/summary:
    get:
      tags: [Admin Reports]
      summary: Revenue and rental summary
      responses:
        '200':
          description: Report summary
  /admin/audit-logs:
    get:
      tags: [Admin Audit]
      summary: List audit logs
      responses:
        '200':
          description: Audit logs
  /admin/users/{id}/role:
    patch:
      tags: [Admin Users]
      summary: Update user role; super admin only
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [customer, admin, super_admin]
      responses:
        '200':
          description: Updated
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100 }
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    User:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        email: { type: string, format: email }
        role:
          type: string
          enum: [customer, admin, super_admin]
    Car:
      type: object
      properties:
        id: { type: integer, format: int64 }
        brand: { type: string }
        model: { type: string }
        year: { type: integer }
        plate_number: { type: string }
        daily_rate: { type: number }
        status:
          type: string
          enum: [available, maintenance, inactive]
        image: { type: string }
    CarInput:
      type: object
      required: [brand, model, year, plate_number, daily_rate]
      properties:
        brand: { type: string }
        model: { type: string }
        year: { type: integer, minimum: 1900 }
        plate_number: { type: string }
        daily_rate: { type: number, minimum: 0 }
        status:
          type: string
          enum: [available, maintenance, inactive]
        image: { type: string }
    Payment:
      type: object
      properties:
        id: { type: integer, format: int64 }
        rental_id: { type: integer, format: int64 }
        amount: { type: number }
        method:
          type: string
          enum: [cash, card, bank_transfer]
        status:
          type: string
          enum: [pending, paid, failed, refunded]
    Rental:
      type: object
      properties:
        id: { type: integer, format: int64 }
        user_id: { type: integer, format: int64 }
        car_id: { type: integer, format: int64 }
        start_date: { type: string, format: date-time }
        end_date: { type: string, format: date-time }
        total_amount: { type: number }
        status:
          type: string
          enum: [requested, approved, rejected, pending_payment, confirmed, active, cancelled, completed]
        payment:
          $ref: '#/components/schemas/Payment'
    AuthEnvelope:
      type: object
      properties:
        data:
          type: object
          properties:
            token: { type: string }
            access_token: { type: string }
            refresh_token: { type: string }
            expires_in: { type: integer }
            user:
              $ref: '#/components/schemas/User'
    CarEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Car'
    RentalEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Rental'
    PaginatedCarsEnvelope:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Car'
            total: { type: integer }
            page: { type: integer }
            page_size: { type: integer }
            total_pages: { type: integer }